feat(cmd): 将主命令迁移为 server 子命令并新增 worker 子命令
- 将 main.go 中的 HTTP 服务器启动逻辑迁移到 server.go 作为 server 子命令
- 新增 worker.go 提供 worker 子命令框架（待实现）
- 简化 main.go 仅保留根命令定义
- 使用 Cobra 子命令模式，便于后续扩展更多命令
