+++ backend/cmd/singer/worker.go
@@ -1,0 +1,23 @@
+package main
+
+import (
+	"github.com/spf13/cobra"
+	"github.com/ygpkg/yg-go/logs"
+)
+
+var workerCmd = &cobra.Command{
+	Use:   "worker",
+	Short: "Start the SingerOS background worker",
+	Long:  `Start the background worker service for processing asynchronous tasks and events.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		logs.Info("Worker service started")
+		
+		// TODO: Implement worker service logic
+		
+		select {}
+	},
+}
+
+func init() {
+	rootCmd.AddCommand(workerCmd)
+}
